+++ lib/project_types/script/layers/application/create_script.rb
@@ -36,7 +36,7 @@ module Script
               )
 
               install_dependencies(ctx, language, script_name, project_creator)
-              script_project_repo.update_or_create_script_json(title: script_name)
+              script_project_repo.update_script_config(title: script_name)
               project
             end
           end