+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 From version 2.6.0, the sections in this file adhere to the [keep a changelog](https://keepachangelog.com/en/1.0.0/) specification.
 ## [Unreleased]
+### Added
+* [#1826](https://github.com/Shopify/shopify-cli/pull/1826): Support using `script.config.yml` file for script configuration
 
 ### Fixed
 * [#1811](https://github.com/Shopify/shopify-cli/pull/1811): Update theme-check to 1.9.0